backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "LangGraph Agent Backend"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
click = ">=8.2.0,<9.0.0"
colorama = ">=0.4.6,<0.5.0"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
httpx = ">=0.28.1,<0.29.0"
numpy = ">=2.2.5,<3.0.0"
pandas = ">=2.2.3,<3.0.0"
dirtyjson = ">=1.0.8,<2.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
wrapt = ">=1.17.2,<2.0.0"
toml = ">=0.10.2,<0.11.0"
tomli = ">=2.2.1,<3.0.0"
packaging = ">=24,<25"
pyproject-toml = ">=0.1.0,<0.2.0"
pyproject-hooks = ">=1.2.0,<2.0.0"
starlette = ">=0.46.2,<0.47.0"
sse-starlette = ">=2.3.5,<3.0.0"
httpcore = ">=1.0.9,<2.0.0"
httptools = ">=0.6.4,<0.7.0"
aiohttp = ">=3.11.18,<4.0.0"
async-timeout = ">=4.0,<5.0"
multidict = ">=6.4.3,<7.0.0"
asgiref = ">=3.8.1,<4.0.0"
httpx-sse = ">=0.4.0,<0.5.0"
python-multipart = ">=0.0.20,<0.0.21"
scipy = ">=1.15.3,<2.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
joblib = ">=1.5.0,<2.0.0"
tqdm = ">=4.67.1,<5.0.0"
mpmath = ">=1.3.0,<2.0.0"
altair = ">=5.5.0,<6.0.0"
sympy = ">=1.14.0,<2.0.0"
torch = ">=2.7.0,<3.0.0"
transformers = ">=4.51.3,<5.0.0"
sentence-transformers = ">=4.1.0,<5.0.0"
huggingface-hub = ">=0.31.2,<0.32.0"
openai = ">=1.78.1,<2.0.0"
anthropic = ">=0.51.0,<0.52.0"
cohere = ">=5.15.0,<6.0.0"
groq = ">=0.24.0,<0.25.0"
tiktoken = ">=0.9.0,<0.10.0"
onnxruntime = ">=1.22.0,<2.0.0"
triton = ">=3.3.0,<4.0.0"
langchain = ">=0.3.25,<0.4.0"
langchain-core = ">=0.3.59,<0.4.0"
langchain-openai = ">=0.3.16,<0.4.0"
langchain-cohere = ">=0.4.4,<0.5.0"
langchain-community = ">=0.3.24,<0.4.0"
langchain-text-splitters = ">=0.3.8,<0.4.0"
langchain-groq = ">=0.3.2,<0.4.0"
langchainhub = ">=0.1.21,<0.2.0"
llama-index = ">=0.12.35,<0.13.0"
llama-index-core = ">=0.12.35,<0.13.0"
llama-index-cli = ">=0.4.1,<0.5.0"
llama-index-agent-openai = ">=0.4.7,<0.5.0"
llama-index-llms-openai = ">=0.3.38,<0.4.0"
llama-index-llms-openai-like = ">=0.3.4,<0.4.0"
llama-index-llms-anthropic = ">=0.6.12,<0.7.0"
llama-index-llms-groq = ">=0.3.1,<0.4.0"
llama-index-llms-ollama = ">=0.5.4,<0.6.0"
llama-index-embeddings-openai = ">=0.3.1,<0.4.0"
llama-index-program-openai = ">=0.3.1,<0.4.0"
llama-index-question-gen-openai = ">=0.3.0,<0.4.0"
llama-index-readers-database = ">=0.4.0,<0.5.0"
llama-index-readers-file = ">=0.4.7,<0.5.0"
llama-index-readers-llama-parse = ">=0.4.0,<0.5.0"
llama-index-readers-web = ">=0.4.1,<0.5.0"
llama-index-indices-managed-llama-cloud = ">=0.6.11,<0.7.0"
llama-parse = ">=0.6.22,<0.7.0"
qdrant-client = ">=1.14.2,<2.0.0"
llama-index-vector-stores-qdrant = ">=0.6.0,<0.7.0"
nltk = ">=3.9.1,<4.0.0"
arxiv = ">=2.2.0,<3.0.0"
bs4 = ">=0.0.2,<0.0.3"
feedparser = ">=6.0.11,<7.0.0"
newspaper3k = ">=0.2.8,<0.3.0"
regex = ">=2024.11.6,<2025.0.0"
soupsieve = ">=2.7,<3.0"
wikipedia = ">=1.4.0,<2.0.0"
s3transfer = ">=0.12.0,<0.13.0"
rsa = ">=4.9.1,<5.0.0"
jmespath = ">=1.0.1,<2.0.0"
google-auth = ">=2.40.1,<3.0.0"
googleapis-common-protos = ">=1.70.0,<2.0.0"
opentelemetry-api = ">=1.33.0,<2.0.0"
opentelemetry-sdk = ">=1.33.0,<2.0.0"
opentelemetry-proto = ">=1.33.0,<2.0.0"
opentelemetry-util-http = ">=0.54b0,<0.55"
opentelemetry-instrumentation = ">=0.54b0,<0.55"
opentelemetry-instrumentation-asgi = ">=0.54b0,<0.55"
opentelemetry-instrumentation-fastapi = ">=0.54b0,<0.55"
opentelemetry-exporter-otlp-proto-common = ">=1.33.0,<2.0.0"
opentelemetry-exporter-otlp-proto-grpc = ">=1.33.0,<2.0.0"
opentelemetry-semantic-conventions = ">=0.54b0,<0.55"
pydantic = ">=2.11.4,<3.0.0"
pydantic-settings = ">=2.9.1,<3.0.0"
build = ">=1.2.2.post1,<2.0.0"
distro = ">=1.9.0,<2.0.0"
sqlalchemy = ">=2.0.40,<3.0.0"
cassandra-driver = ">=3.29.2,<4.0.0"
cassio = ">=0.1.10,<0.2.0"
tokenizers = ">=0.21.1,<0.22.0"
pypdf = ">=5.5.0,<6.0.0"
pypdf2 = ">=3.0.1,<4.0.0"
langgraph = ">=0.4.3,<0.5.0"
boto3 = "^1.38.18"
ruff = "^0.11.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks.format]
sequence = ["ruff-fix", "isort", "black"]

[tool.poe.tasks.ruff-fix]
cmd = "ruff check . --fix"

[tool.poe.tasks.isort]
cmd = "isort ."

[tool.poe.tasks.black]
cmd = "black ."